Move schema validation tests.
