+++ tests/ludwig/config_validation/test_validate_config_features.py
@@ -1,7 +1,7 @@
 import pytest
 from jsonschema.exceptions import ValidationError
 
-from ludwig.schema import validate_config
+from ludwig.config_validation import validate_config
 from tests.integration_tests.utils import binary_feature, category_feature, number_feature, text_feature
 
 
@@ -14,7 +14,7 @@ def test_config_input_output_features():
         "output_features": [binary_feature(decoder={"type": "regressor"})],
     }
 
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
 
 def test_incorrect_input_features_config():
@@ -28,7 +28,7 @@ def test_incorrect_input_features_config():
     # TODO(ksbrar): Circle back after discussing whether additional properties should be allowed long-term.
     # # Not a preprocessing param for category feature
     # with pytest.raises(ValidationError):
-    #     validate_config(config)
+    #     validate_config(config, include_auxiliary_validations=False)
 
     config = {
         "input_features": [
@@ -39,7 +39,7 @@ def test_incorrect_input_features_config():
 
     # Incorrect type for padding_symbol preprocessing param
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
     config = {
         "input_features": [
@@ -51,7 +51,7 @@ def test_incorrect_input_features_config():
 
     # Incorrect type for padding_symbol preprocessing param
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
 
 def test_incorrect_output_features_config():
@@ -64,4 +64,4 @@ def test_incorrect_output_features_config():
 
     # Invalid decoder for binary output feature
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
