+++ tests/ludwig/config_validation/test_validate_config_misc.py
@@ -19,7 +19,8 @@ from ludwig.constants import (
     TYPE,
 )
 from ludwig.features.feature_registries import output_type_registry
-from ludwig.schema import get_schema, validate_config
+from ludwig.schema import get_schema
+from ludwig.config_validation import validate_config
 from ludwig.schema.defaults.defaults import DefaultsConfig
 from ludwig.schema.features.preprocessing.audio import AudioPreprocessingConfig
 from ludwig.schema.features.preprocessing.bag import BagPreprocessingConfig
@@ -84,12 +85,12 @@ def test_config_features():
         "input_features": all_input_features,
         "output_features": all_output_features,
     }
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
     # make sure all defaults provided also registers as valid
 
     config = ModelConfig.from_dict(config).to_dict()
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
     # test various invalid output features
     input_only_features = [
@@ -103,7 +104,7 @@ def test_config_features():
 
         dtype = input_feature["type"]
         with pytest.raises(ValidationError, match=rf"^'{dtype}' is not one of .*"):
-            validate_config(config)
+            validate_config(config, include_auxiliary_validations=False)
 
 
 def test_config_encoders():
@@ -116,7 +117,7 @@ def test_config_encoders():
             "output_features": [category_feature(decoder={"type": "classifier", "vocab_size": 2}, reduce_input="sum")],
             "combiner": {"type": "concat", "output_size": 14},
         }
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
 
 def test_config_tabnet():
@@ -153,7 +154,7 @@ def test_config_tabnet():
             "validation_field": "label",
         },
     }
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
 
 def test_config_bad_feature_type():
@@ -164,7 +165,7 @@ def test_config_bad_feature_type():
     }
 
     with pytest.raises(ValidationError, match=r"^'fake' is not one of .*"):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
 
 def test_config_bad_encoder_name():
@@ -175,7 +176,7 @@ def test_config_bad_encoder_name():
     }
 
     with pytest.raises(ValidationError, match=r"^'fake' is not one of .*"):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
 
 # TODO(ksbrar): Circle back after discussing whether additional properties should be allowed long-term.
@@ -199,7 +200,7 @@ def test_config_bad_encoder_name():
 #     }
 
 #     with pytest.raises(ValidationError, match=r"^Additional properties are not allowed .*"):
-#         validate_config(config)
+#         validate_config(config, include_auxiliary_validations=False)
 
 
 def test_config_fill_values():
@@ -212,7 +213,7 @@ def test_config_fill_values():
             ],
             "output_features": [binary_feature(preprocessing={"fill_value": binary_fill_value})],
         }
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
     bad_vector_fill_values = ["one two three", "1,2,3", 0]
     bad_binary_fill_values = ["one", 2, "maybe"]
@@ -224,7 +225,7 @@ def test_config_fill_values():
             "output_features": [binary_feature(preprocessing={"fill_value": binary_fill_value})],
         }
         with pytest.raises(ValidationError):
-            validate_config(config)
+            validate_config(config, include_auxiliary_validations=False)
 
 
 def test_validate_with_preprocessing_defaults():
@@ -263,9 +264,9 @@ def test_validate_with_preprocessing_defaults():
         },
     }
 
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
     config = ModelConfig.from_dict(config).to_dict()
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
 
 def test_defaults_schema():
@@ -315,12 +316,12 @@ def test_validate_defaults_schema():
         },
     }
 
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
     config[DEFAULTS][CATEGORY][NAME] = "TEST"
 
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
 
 def test_validate_no_trainer_type():
@@ -335,23 +336,23 @@ def test_validate_no_trainer_type():
     }
 
     # Ensure validation succeeds with ECD trainer params and ECD model type
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
     # Ensure validation fails with ECD trainer params and GBM model type
     config[MODEL_TYPE] = MODEL_GBM
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
     # Switch to trainer with valid GBM params
     config[TRAINER] = {"tree_learner": "serial"}
 
     # Ensure validation succeeds with GBM trainer params and GBM model type
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
     # Ensure validation fails with GBM trainer params and ECD model type
     config[MODEL_TYPE] = MODEL_ECD
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
 
 def test_schema_no_duplicates():
