+++ tests/ludwig/config_validation/test_validate_config_trainer.py
@@ -2,7 +2,7 @@ import pytest
 from jsonschema.exceptions import ValidationError
 
 from ludwig.constants import TRAINER
-from ludwig.schema import validate_config
+from ludwig.config_validation import validate_config
 from ludwig.schema.optimizers import optimizer_registry
 from ludwig.schema.trainer import ECDTrainerConfig
 from tests.integration_tests.utils import binary_feature, category_feature, number_feature
@@ -23,14 +23,14 @@ def test_config_trainer_empty_null_and_default():
         },
         TRAINER: {},
     }
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
     config[TRAINER] = None
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
     config[TRAINER] = ECDTrainerConfig.Schema().dump({})
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
 
 def test_config_trainer_bad_optimizer():
@@ -45,29 +45,29 @@ def test_config_trainer_bad_optimizer():
         },
         TRAINER: {},
     }
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
     # Test manually set-to-null optimizer vs unspecified:
     config[TRAINER]["optimizer"] = None
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
     assert ECDTrainerConfig.Schema().load({}).optimizer is not None
 
     # Test all types in optimizer_registry supported:
     for key in optimizer_registry.keys():
         config[TRAINER]["optimizer"] = {"type": key}
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
     # Test invalid optimizer type:
     config[TRAINER]["optimizer"] = {"type": 0}
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
     config[TRAINER]["optimizer"] = {"type": {}}
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
     config[TRAINER]["optimizer"] = {"type": "invalid"}
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
 
 def test_optimizer_property_validation():
@@ -82,30 +82,30 @@ def test_optimizer_property_validation():
         },
         TRAINER: {},
     }
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
     # Test that an optimizer's property types are enforced:
     config[TRAINER]["optimizer"] = {"type": "rmsprop"}
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
     config[TRAINER]["optimizer"]["momentum"] = "invalid"
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
     # Test extra keys are excluded and defaults are loaded appropriately:
     config[TRAINER]["optimizer"]["momentum"] = 10
     config[TRAINER]["optimizer"]["extra_key"] = "invalid"
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
     assert not hasattr(ECDTrainerConfig.Schema().load(config[TRAINER]).optimizer, "extra_key")
 
     # Test bad parameter range:
     config[TRAINER]["optimizer"] = {"type": "rmsprop", "eps": -1}
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
     # Test config validation for tuple types:
     config[TRAINER]["optimizer"] = {"type": "adam", "betas": (0.1, 0.1)}
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
 
 def test_clipper_property_validation():
@@ -120,13 +120,13 @@ def test_clipper_property_validation():
         },
         TRAINER: {},
     }
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
 
     # Test null/empty clipper:
     config[TRAINER]["gradient_clipping"] = None
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
     config[TRAINER]["gradient_clipping"] = {}
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
     assert (
         ECDTrainerConfig.Schema().load(config[TRAINER]).gradient_clipping
         == ECDTrainerConfig.Schema().load({}).gradient_clipping
@@ -135,22 +135,22 @@ def test_clipper_property_validation():
     # Test invalid clipper type:
     config[TRAINER]["gradient_clipping"] = 0
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
     config[TRAINER]["gradient_clipping"] = "invalid"
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
     # Test that an optimizer's property types are enforced:
     config[TRAINER]["gradient_clipping"] = {"clipglobalnorm": None}
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
     config[TRAINER]["gradient_clipping"] = {"clipglobalnorm": 1}
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
     config[TRAINER]["gradient_clipping"] = {"clipglobalnorm": "invalid"}
     with pytest.raises(ValidationError):
-        validate_config(config)
+        validate_config(config, include_auxiliary_validations=False)
 
     # Test extra keys are excluded and defaults are loaded appropriately:
     config[TRAINER]["gradient_clipping"] = {"clipnorm": 1}
     config[TRAINER]["gradient_clipping"]["extra_key"] = "invalid"
-    validate_config(config)
+    validate_config(config, include_auxiliary_validations=False)
     assert not hasattr(ECDTrainerConfig.Schema().load(config[TRAINER]).gradient_clipping, "extra_key")
